Move SportPaper events outside of sport package
