+++ util/src/main/java/tc/oc/pgm/util/event/entity/ExplosionPrimeByEntityEvent.java
@@ -1,4 +1,4 @@
-package tc.oc.pgm.util.event.sport.entity;
+package tc.oc.pgm.util.event.entity;
 
 import static com.google.common.base.Preconditions.checkNotNull;
