+++ util/src/main/java/tc/oc/pgm/util/event/player/PlayerSpawnEntityEvent.java
@@ -1,4 +1,4 @@
-package tc.oc.pgm.util.event.sport.player;
+package tc.oc.pgm.util.event.player;
 
 import org.bukkit.Location;
 import org.bukkit.entity.Entity;
