+++ util/src/main/java/tc/oc/pgm/util/tablist/PlayerTabEntry.java
@@ -6,7 +6,7 @@ import java.util.UUID;
 import javax.annotation.Nullable;
 import net.kyori.adventure.text.Component;
 import org.bukkit.entity.Player;
-import tc.oc.pgm.util.event.sport.player.PlayerSkinPartsChangeEvent;
+import tc.oc.pgm.util.event.player.PlayerSkinPartsChangeEvent;
 import tc.oc.pgm.util.named.NameStyle;
 import tc.oc.pgm.util.nms.NMSHacks;
 import tc.oc.pgm.util.skin.Skin;
